requirements.txt
alembic>=1.10.0
asyncpg>=0.28.0
email-validator>=2.0.0
fastapi>=0.115.0
passlib[argon2]>=1.7.4
psycopg2-binary>=2.9.5
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
returns>=0.20.0
sqlalchemy[asyncio]>=2.0.0
uvicorn>=0.34.0
